+++ c/chunk.h
@@ -95,14 +95,14 @@ typedef enum {
 typedef struct {
     int count;
     int capacity;
-    uint8_t *code;
+    uint16_t *code;
     int *lines;
     ValueArray constants;
 } Chunk;
 
 void initChunk(Chunk *chunk);
 void freeChunk(VM *vm, Chunk *chunk);
-void writeChunk(VM *vm, Chunk *chunk, uint8_t byte, int line);
+void writeChunk(VM *vm, Chunk *chunk, uint16_t byte, int line);
 int addConstant(VM *vm, Chunk *chunk, Value value);
 
 #endif //C_CHUNK_H
